+++ src/DiscreteSignals.h
@@ -206,13 +206,13 @@ private:
 
 private:
     std::string m_sInternalModuleName;
-    uint8_t m_uiInternalModuleId;
+    uint16_t m_uiInternalModuleId;
 
     std::string m_sDeviceControlName;
-    uint8_t m_uiDeviceControlId;
+    uint16_t m_uiDeviceControlId;
 
     std::string m_sDataStoreName;
-    uint8_t m_uiDataStoreId;
+    uint16_t m_uiDataStoreId;
 
     std::vector<CTaskInterface*> m_vpxAlarmHandlers;
     std::vector<CTaskInterface*>::iterator m_xAlarmHandlersIterator;
